Fixed types, indents from code review
