+++ src/app/publisher/dashboard/dashboard.component.html
@@ -52,8 +52,8 @@
           <div class="chart-stats">
             <p
               class="
-              chart-name
-              adsh-copy"
+                chart-name
+                adsh-copy"
             >
               {{ chartSeriesType }}:
             </p>
